forgot to rename main. likely to change again but for now it's good
